refactor: move icecast-metadata-player out of player
