+++ src/demo/src/Player/Player.jsx
@@ -1,81 +1,21 @@
-import React, {
-  useLayoutEffect,
-  useEffect,
-  useState,
-  useCallback,
-} from "react";
+import React, { useLayoutEffect, useState } from "react";
 import AudioSpectrum from "react-audio-spectrum";
-import IcecastMetadataPlayer from "icecast-metadata-player";
 import { ReactComponent as Play } from "./play.svg";
 import { ReactComponent as Pause } from "./pause.svg";
 import styles from "./Player.module.css";
 
-const SELECT_STATION = "Select a station";
-const SELECT_OR_PLAY = "Select a station or press play";
-const LOADING = "Loading...";
 const VISIT_STATION = "Visit this station at ";
 const ICECAST_METADATA_JS_DEMO = "Icecast Metadata JS Demo";
 
-const Player = ({ station }) => {
-  const [audioElement] = useState(new Audio());
+const Player = ({ station, playing, toggle, audioElement, metadata }) => {
   const [[audioHeight, audioWidth], setSpectrumSize] = useState([0, 0]);
   const [meters, setMeters] = useState(0);
 
-  const [metadata, setMetadata] = useState(SELECT_STATION);
-  const [icecast, setIcecast] = useState();
-  const [playing, setPlaying] = useState(false);
-
-  // begin playing when a new station is selected
-  useEffect(() => {
-    if (icecast) {
-      icecast.play();
-      // browser audio element integration
-      audioElement.addEventListener("pause", icecast.stop);
-      return () => audioElement.removeEventListener("pause", icecast.stop);
-    }
-  }, [audioElement, icecast]);
-
-  // change station
-  useEffect(() => {
-    if (station) {
-      icecast && icecast.stop();
-
-      setIcecast(
-        new IcecastMetadataPlayer(station.endpoint, {
-          onMetadata: (meta) => {
-            console.log(meta);
-            setMetadata(meta);
-          },
-          onStop: () => {
-            setPlaying(false)
-            setMetadata(SELECT_OR_PLAY);
-          },
-          onLoading: () => {
-            setPlaying(true)
-            setMetadata(LOADING);
-          },
-          onPlay: () => {
-            setPlaying(true)
-          },
-          icyDetectionTimeout: 5000,
-          metadataTypes: station.metadataTypes,
-          audioElement,
-        })
-      );
-    }
-  }, [station, audioElement]);
-
   // update metadata in title
-  useEffect(() => {
-    const title = metadata.StreamTitle || metadata.TITLE;
-    document.title = title
-      ? `${title} | ${ICECAST_METADATA_JS_DEMO}`
-      : ICECAST_METADATA_JS_DEMO;
-  }, [metadata]);
-
-  const toggle = useCallback(() => {
-    playing ? icecast.stop() : icecast.play();
-  }, [icecast, playing]);
+  const title = metadata.StreamTitle || metadata.TITLE;
+  document.title = title
+    ? `${title} | ${ICECAST_METADATA_JS_DEMO}`
+    : ICECAST_METADATA_JS_DEMO;
 
   useLayoutEffect(() => {
     const updateSize = () => {
